Read pGlyco3-pGlycoQuant result

If you used pGlyco3 for intact glycopeptide identification, and used pGlycoQuant for quantification, this is the function for you. It reads in a pGlycoQuant result file and returns a glyexp::experiment() object. Both label-free and TMT quantification are supported.

Usage

read_pglyco3_pglycoquant(
  fp,
  sample_info_fp = NULL,
  name = NULL,
  quant_method,
  tmt_type = NULL,
  ref_channel = NULL,
  parse_structure = FALSE,
  quant_aggr_method = c("gf", "gp", "gfs", "gps", "det"),
  glycan_type = c("N", "O")
)

Arguments

fp

File path of the pGlyco3 result file.

sample_info_fp

File path of the sample information file.

name

Name of the experiment. If not provided, a default name with current time will be used.

quant_method

Quantification method. Either "label-free" or "TMT".

tmt_type

TMT type. Required if quant_method is "TMT". Options are "TMT-6plex", "TMT-10plex", "TMT-11plex", "TMTpro-10plex", "TMTpro-16plex", "TMTpro-18plex".

ref_channel

Reference channel for TMT normalization. A character scalar. Required if quant_method is "TMT". The reference channel must be one of the valid TMT channels.

TMT-6plex: 126, 127, 128, 129, 130, 131.

TMT-10plex: 126, 127N, 127C, 128N, 128C, 129N, 129C, 130N, 130C, 131.

TMT-11plex: 126, 127N, 127C, 128N, 128C, 129N, 129C, 130N, 130C, 131N, 131C.

TMTpro-10plex: 126, 127N, 128N, 129N, 130N, 131N, 132N, 133N, 134N, 135N.

TMTpro-16plex: 126, 127N, 127C, 128N, 128C, 129N, 129C, 130N, 130C, 131N, 131C, 132N, 132C, 133N, 133C, 134N.

TMTpro-18plex: 126, 127N, 127C, 128N, 128C, 129N, 129C, 130N, 130C, 131N, 131C, 132N, 132C, 133N, 133C, 134N, 134C, 135N.

parse_structure

Whether to parse glycan structures. Default is FALSE. It can only be set TRUE when quant_aggr_method is "gfs", "gps", or "det". In other cases, the structure column is dropped after aggregation. See the details section and quant_aggr_method argument for more information.

quant_aggr_method

Aggregation method for quantification results.

"gf": Glycoform level. The default. Aggregates all PSMs of the same combination of protein, protein site, and glycan composition.

"gp": Glycopeptide level. Aggregates all PSMs of the same combination of peptide, peptide site, and glycan composition.

"gfs": Glycoform level with glycan structure. Aggregates all PSMs of the same combination of protein, protein site, and glycan structure.

"gps": Glycopeptide level with glycan structure. Aggregates all PSMs of the same combination of peptide, peptide site, and glycan structure.

"det": Detailed level. Aggregates all PSMs of the same combination of peptide, protein, gene, glycan composition, glycan structure, peptide site, protein site, charge, and modifications.

glycan_type

Glycan type. Either "N" or "O". Default is "N".

Value

An glyexp::experiment() object.

Input

You should use the "Quant.spectra.list" file in the pGlycoQuant result folder. Files from pGlyco3 result folder are not needed. For instructions on how to use pGlyco3 and pGlycoQuant, please refer to the manual: pGlycoQuant.

The sample information file should be a csv file with the first column named sample, and the rest of the columns being sample information. The sample column must match the RawName column in the pGlyco3 result file, although the order can be different.

For results from TMT quantification, the sample information file should have two additional columns besides sample: raw_name and channel. raw_name refers to the raw file names, matching the RawName column in the pGlyco3 result file. channel refers to the TMT channel names. For valid TMT channels, see the ref_channel argument.

You can put any useful information in the sample information file. Recommended columns are:

group: grouping or conditions, e.g. "control" or "tumor", required for most downstream analyses

batch: batch information, required for batch effect correction

bio_replicate: Indicates the identity of biologically distinct samples. While the sample column is typically used as a unique identifier for each run, bio_replicate specifies whether the samples originate from the same biological source. In common experimental designs, multiple technical replicates are often derived from the same biological sample, making the bio_replicate column essential for distinguishing biological replicates from technical replicates.

Output

This function returns a glyexp::experiment() object.

If parse_structure is TRUE, a list of parsed glycan_graph objects will be stored in the experiment, with the glycan structure strings as names. It can be accessed by exp$glycan_graphs or exp[["glycan_graphs"]]. Although the results from pGlyco3 provide structural information, many of these structures are speculative, and pGlyco3 doesn’t include any quality control for the structural speculation process. If you choose to enable this option, please interpret the analysis results with caution. This is why parse_structure is FALSE by default for this function.

The following columns could be found in the variable information tibble:

charge: integer, charge state

peptide: character, peptide sequence

modifications: character, modifications other than glycosylation, separated by semicolon, e.g. 5,Carbamidomethyl[C];10,Carbamidomethyl[C]

glycan_composition: character, glycan composition, e.g. "H5N4F1A1"

glycan_structure: character, pGlyco-style structure strings, renamed from PlausibleStruct in the original file

peptide_site: integer, site of glycosylation on peptide

proteins: character, protein names, separated by semicolon

genes: character, gene names, separated by semicolon

protein_sites: character, site of glycosylation on protein, separated by semicolon

Some of them might not be present, depending on the aggregation method. See quant_aggr_method argument for more information.

Aggregation

pGlycoQuant quantifies glycopeptides at the PSM level. Therefore, a glycopeptide could have multiple rows in the result file, due to different charge states, modifications, or just duplicate identifications. In this regard, we need to aggregate the quantification results to the level we are interested in. This function provides several aggregation methods, specified by the quant_aggr_method argument. The most commonly used method is "gf" (glycoform), which aggregates all PSMs of the same combination of protein, protein site, and glycan composition. The word "aggregate" here means "sum" for label-free quantification, and "take the median" for TMT quantification. For other aggregation methods, see the quant_aggr_method argument.