Read pGlyco3-pGlycoQuant result

If you used pGlyco3 for intact glycopeptide identification, and used pGlycoQuant for quantification, this is the function for you. It reads in a pGlycoQuant result file and returns a glyexp::experiment() object. Currently only label-free quantification is supported.

Usage

read_pglyco3_pglycoquant(
  fp,
  sample_info_fp = NULL,
  name = NULL,
  quant_method = c("label-free", "TMT"),
  parse_structure = FALSE,
  glycan_type = c("N", "O")
)

Arguments

fp

File path of the pGlyco3 result file.

sample_info_fp

File path of the sample information file.

name

Name of the experiment. If not provided, a default name with current time will be used.

quant_method

Quantification method. Either "label-free" or "TMT".

parse_structure

Whether to parse glycan structures. Default is FALSE.

glycan_type

Glycan type. Either "N" or "O". Default is "N".

Value

An glyexp::experiment() object.

Input

You should use the "Quant.spectra.list" file in the pGlycoQuant result folder. Files from pGlyco3 result folder are not needed. For instructions on how to use pGlyco3 and pGlycoQuant, please refer to the manual: pGlycoQuant.

The sample information file should be a csv file with the first column named sample, and the rest of the columns being sample information. The sample column must match the RawName column in the pGlyco3 result file, although the order can be different.

You can put any useful information in the sample information file. Recommended columns are:

group: grouping or conditions, e.g. "control" or "tumor", required for most downstream analyses

batch: batch information, required for batch effect correction

bio_replicate: Indicates the identity of biologically distinct samples. While the sample column is typically used as a unique identifier for each run, bio_replicate specifies whether the samples originate from the same biological source. In common experimental designs, multiple technical replicates are often derived from the same biological sample, making the bio_replicate column essential for distinguishing biological replicates from technical replicates.

Output

This function returns a glyexp::experiment() object.

If parse_structure is TRUE, a list of parsed glycan_graph objects will be stored in the experiment, with the glycan structure strings as names. It can be accessed by exp$glycan_graphs or exp[["glycan_graphs"]]. Although the results from pGlyco3 provide structural information, many of these structures are speculative, and pGlyco3 doesn’t include any quality control for the structural speculation process. If you choose to enable this option, please interpret the analysis results with caution. This is why parse_structure is FALSE by default for this function.

The following columns could be found in the variable information tibble:

charge: integer, charge state

peptide: character, peptide sequence

modifications: character, modifications other than glycosylation, separated by semicolon, e.g. 5,Carbamidomethyl[C];10,Carbamidomethyl[C]

glycan_composition: character, glycan composition, e.g. "H5N4F1A1"

glycan_structure: character, pGlyco-style structure strings, renamed from PlausibleStruct in the original file

peptide_site: integer, site of glycosylation on peptide

proteins: character, protein names, separated by semicolon

genes: character, gene names, separated by semicolon

protein_sites: character, site of glycosylation on protein, separated by semicolon